seems to work

+++ midgard/xml/export_common.cc
@@ -1,4 +1,4 @@
-// $Id: export_common.cc,v 1.3 2001/12/16 21:47:49 christof Exp $
+// $Id: export_common.cc,v 1.4 2001/12/17 09:53:44 christof Exp $
 /*  Midgard Roleplaying Character Generator
  *  Copyright (C) 2001 Christof Petig
  *
@@ -71,9 +71,9 @@ int fetch_and_write_int_attrib(FetchIStream &is,std::ostream &o,const std::strin
 }
 
 double fetch_float(FetchIStream &is,double standard=0)
-{  double val;
+{  float val;
    
-   is >> FetchIStream::MapNull<double>(val,standard);
+   is >> FetchIStream::MapNull<float>(val,standard);
    return val;
 }
